댓글기능 반정도 완성

+++ src/main/java/com/platform/learning/reactbootproject/editor/communitydto/CommunityCommentDTO.java
@@ -9,5 +9,5 @@ public class CommunityCommentDTO {
     private int community_comment_community_pk;
     private String community_comment_content;
     private String community_comment_date;
-    private Long community_comment_isgood;
+    private int community_comment_isgood;
 }
